fix: Move candidate-scan pins out of the quotes module's tests
The two new tests measure into coverage only through the platform-dependent
in-module-test filter, which is the same phantom-patch-line trap #1365 hit;
relocating them to parser/src/tests (excluded when the lcov report is
generated, on every platform) makes the measurement deterministic. The
differential reference stays in the quotes module as a #[cfg(test)] item so
it keeps private access, with its lines measured symmetrically everywhere.
